yolov5-master/requirements.txt
absl-py==0.13.0
altgraph==0.17
asgiref==3.4.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
Cython==0.29.26
Django==3.2.7
django-cors-headers==3.10.0
djangorestframework==3.12.4
docopt==0.6.2
easydict==1.9
filelock==3.4.2
flake8==4.0.1
future==0.18.2
gdown==4.2.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
h5py==3.6.0
idna==3.2
imageio==2.13.5
imutils==0.5.4
isort==4.3.21
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mccabe==0.6.1
mysql-connector-python==8.0.27
numpy==1.21.1
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.5.62
opencv-python-headless==3.4.17.61
paho-mqtt==1.6.1
pandas==1.3.5
pefile==2021.5.24
Pillow==9.0.0
pipreqs==0.4.11
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tb-nightly==2.8.0a20220105
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.25.8
Werkzeug==2.0.1
yacs==0.1.8
yapf==0.32.0
yarg==0.1.9
